feat(elections): adding blinded proposals
Implements blind voting to improve the Andromeda consensus mechanism.
Instead of listening for proposals at a given block number, there is now a voting period and a reveal period defined in seconds.
Withing the voting period operators can submit blinded proposals which are generated by using a random secret.
All operators will submit unique hashes to avoid looking at other people's votes.
When the voting period is over, operators reveal the raw proposal and the secret used to generate the blind proposal.

Once all votes have been revealed, the normal count process takes places.
Currently, count is a Proof of Authority mechanism that allows registry operators to participate in consensus.
All raw proposals are counted and the proposal with the most votes is claimed as the winning proposal.

Block explorers may then use the winning proposal to verify validity of the data which is queried from the winning operators.

ENG-325
